pyproject.toml
[tool.poetry]
name = "meal-calorie-api"
version = "0.1.0"
description = "FastAPI + Postgres: Meal Calorie Count Generator"
authors = ["Rahul Kumar <therahulkumar40.com>"]
readme = "README.md"
packages = [{ include = "app" }]
# require Python 3.11+
[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.115"
pydantic-settings = ">=2.2"
httpx = ">=0.27"
rapidfuzz = ">=3.7"
SQLAlchemy = ">=2.0"
psycopg = {version = ">=3.1", extras = ["binary"]}
alembic = ">=1.13"
passlib = {version = "1.7.4", extras = ["bcrypt"]}
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
email-validator = ">=2.1"
slowapi = ">=0.1.9"
python-multipart = "^0.0.9"
pydantic = ">=2.4"
bcrypt = "4.0.1"
pyjwt = ">=2.8"
cachetools = ">=5.3"
uvicorn = "0.27.1"
click = "8.1.7"
typer = "0.12.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.8"
httpx = {version = "^0.27.0", extras = ["http2"]}
coverage = "^7.6.1"
ruff = "^0.5.5"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.0"
types-passlib = "^1.7.7.20240327"
types-python-jose = "^3.3.4.20240106"
pytest-cov = "^7.0.0"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["app", "tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
